Add files via upload
Here are my Python code for cleaning and my SQL query to build the tables.  Also my final data for the tableau dashboard.  

The project ended up displaying the percentage of contractors that every provider provides in the state and the nursing minutes per resident in the top 50 most populous metro areas.  

Last is the final data set (Nursing Edited 2) as long with the two initial data sets.